+++ lib/src/commands/types/mp_command_description_type.dart
@@ -72,6 +72,7 @@ enum MPCommandDescriptionType {
   simplifyToStraight,
   splitLineAtSelectedPoints,
   splitLinesAtCrossings,
+  toggleBorderOption,
   toggleReverseOption,
   toggleSmoothOption;